Rename awsIamAuthorizerProcessor for consistency
