+++ src/contrib/awsIamAuthorizerProcessor/index.test.ts
